try to implement email verify (note note complited)

+++ renters_management_back_end/database/migrations/2024_05_11_002855_create_rent_payments.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->string('year')->nullable(false);
             $table->enum('month',['1','2','3','4','5','6','7','8','9','10','11','12'])->nullable(false);
             $table->enum('state',['payed','partially_payed','not_payed'])->nullable(false);
-            $table->float('payed_amount')->default(0);
-            $table->float('remain_amount');
+            $table->decimal('payed_amount',16,2)->default(0);
+            $table->decimal('remain_amount',16,2);
             $table->softDeletes();
             $table->timestamps();
